+++ src/mainOLD.cpp
@@ -31,7 +31,7 @@ int main(int argc, char **argv){
     float* cpu_buffer = chw_data.data();
 
     // double blob_from_image_sum = 0;
-    // for (size_t i = 0; i < LOOP_COUNT; i++)
+    // for (int i = 0; i < LOOP_COUNT; i++)
     // {
     //     cv::Mat img(H,W,CV_32FC3,cpu_buffer);
     //     double t0 = now_ms();
@@ -46,7 +46,7 @@ int main(int argc, char **argv){
     // float* gpu_buffer;
     // cudaMalloc((void**)&gpu_buffer, H * W * C * sizeof(float));
     // double CUDA_CPU_to_GPU_sum = 0;
-    // for (size_t i = 0; i < LOOP_COUNT; i++)
+    // for (int i = 0; i < LOOP_COUNT; i++)
     // {
     //     double t0 = now_ms();
     //     cudaMemcpyAsync(gpu_buffer,cpu_buffer,H * W * C * sizeof(float),cudaMemcpyHostToDevice);
@@ -60,7 +60,7 @@ int main(int argc, char **argv){
 
 
     // double CUDA_GPU_to_CPU_sum = 0;
-    // for (size_t i = 0; i < LOOP_COUNT; i++)
+    // for (int i = 0; i < LOOP_COUNT; i++)
     // {
     //     double t0 = now_ms();
     //     cudaMemcpyAsync(cpu_buffer,gpu_buffer,H * W * C * sizeof(float),cudaMemcpyDeviceToHost);
@@ -74,7 +74,7 @@ int main(int argc, char **argv){
     auto prosser = std::make_unique<CudaFun>(H,W,C);
     double min_time = std::numeric_limits<double>::max();
     // double CHW_2_1HWC_sum = 0;
-    for (size_t i = 0; i < LOOP_COUNT; i++)
+    for (int i = 0; i < LOOP_COUNT; i++)
     {
         double t0 = now_ms();
         prosser->Data2Image(&cpu_buffer,H*W*C,1.0f/255.0f);
@@ -91,7 +91,7 @@ int main(int argc, char **argv){
 
 
     // double HWC_2_1CHW_sum = 0;
-    // for (size_t i = 0; i < LOOP_COUNT; i++)
+    // for (int i = 0; i < LOOP_COUNT; i++)
     // {
     //     double t0 = now_ms();
     //     CUDA_1HWC_2_1CHW(&cpu_buffer,H,W,C,1.0f/255.0f);
